translator/requirements/requirements_prod.txt
fastapi>=0.95.0,<0.96.0
uvicorn[standard]>=0.21.0,<0.22.0
pydantic~=1.10.13
PyYAML~=6.0.1
colorama~=0.4.6
